+++ coordinates.cpp
@@ -1,4 +1,4 @@
 #include<coordinates.h>
 Coordinates::Coordinates() = default;
-Coordinates::Coordinates(size_t x, size_t y) : x(x), y(y) {};
+Coordinates::Coordinates(int x, int y) : x(x), y(y) {};
